Rename header file
