+++ chromium_src/chrome/browser/component_updater/widevine_cdm_component_installer.cc
@@ -20,7 +20,7 @@
 
 #include "base/check.h"
 #include "base/functional/callback.h"
-#include "brave/browser/widevine/buildflags.h"
+#include "brave/browser/widevine/static_buildflags.h"
 #include "brave/browser/widevine/widevine_utils.h"
 #include "components/component_updater/component_updater_service.h"
