+++ chromium_src/components/update_client/DEPS
@@ -1,5 +1,5 @@
 include_rules = [
   "+brave/components/constants",
   "+brave/browser/widevine/constants.h",
-  "+brave/browser/widevine/buildflags.h",
+  "+brave/browser/widevine/static_buildflags.h",
 ]
